+++ application/home/model/teacherModel.class.php
@@ -130,11 +130,18 @@ class teacherModel extends model{
 	 */
 	public function deleteWorkById(){
 		$id = (int)$_GET['id'];
-		$sql = "delete from `teacher_work` where id=:id";
-		//通过预处理执行SQL
-		$this->db->execute($sql,array(':id'=>$id),$flag);
-		//返回是否执行成功
-		return $flag;
+		
+		$sql = "select `address` from `teacher_work` where id=$id";
+		$data = $this->db->fetchRow($sql);
+	
+		if(unlink($data['address'])){					
+			$sql = "delete from `teacher_work` where id=:id";
+			//通过预处理执行SQL
+			$this->db->execute($sql,array(':id'=>$id),$flag);
+		
+			//返回是否执行成功
+			return $flag;				
+		}
 	}
 	/*
 	 * 获得学生列表
@@ -226,11 +233,16 @@ class teacherModel extends model{
 	 */
 	public function deleteMaterialById(){
 		$id = (int)$_GET['id'];
-		$sql = "delete from `course_material` where id=:id";
-		//通过预处理执行SQL
-		$this->db->execute($sql,array(':id'=>$id),$flag);
-		//返回是否执行成功
-		return $flag;
+		$sql = "select `address` from `course_material` where id=$id";
+		$data = $this->db->fetchRow($sql);
+	
+		if(unlink($data['address'])){
+			$sql = "delete from `course_material` where id=:id";
+			//通过预处理执行SQL
+			$this->db->execute($sql,array(':id'=>$id),$flag);
+			//返回是否执行成功
+			return $flag;
+		}	
 	}
 	/**
 	 * 资料下载
